requirements.txt
transformers
scikit-learn
pillow
torch
wget
accelerate